+++ app/src/main/java/com/example/foodwatch/EditMealFragment.kt
@@ -197,7 +197,7 @@ class EditMealFragment : Fragment(R.layout.fragment_editmeal) {
                 ingredientViewModel.addOrUpdateIngredients(ingredient).join()
                 val ingredientId = ingredientViewModel.getIngredientIdByName(ingredient).await()
                 if(ingredientId != null){
-                    ingredientIds.add(ingredientId)
+                    ingredientIds.add(ingredientId.toInt())
                 }
                 else{
                     Toast.makeText(requireContext(), "Error finding ingredientID!", Toast.LENGTH_LONG).show()
